More control for isa_plugin() arguments
When a plugin is wrapped with an isa_plugin() call, the plugin we're
inheriting from will receive the same arguments as the wrapper. This
patch allows to modify the argument list for the wrapped plugin, i.e.
 * delete arguments only valid for the wrapping plugin
 * add additional arguments
Modifying @_ directly only allows something like "for (@_) { s/a/b/g }",
not adding or deleting. Returning the new @_ will much likely break
existing plugins.